Create MEDICAL INVENTORY MANAGEMENT
MediStock is an inventory management system which manages the medicines
of the distribution company. This company dlistributes the medicines outside to
a large number of medical stores. This company is managing the stock by
making use of the paper based system. The pharmacists makes the orders
through medical sales representative. Medical sales representative orders the
medicinəs by contacting the company through the phone. Company staff writes
the pharmacist' orders on the book. Hence the company is having sifficulties in
managing the stock when they are using the paper basedsystem.
Medical inventory management is a critical component of healthcare operations,
as it ensures timely patient care, reduces costs, and improves profitability.